Move NextNodeBlock out of Question namespace
